HUGE amounts of refactoring + getting new file IDs
