+++ bot.py
@@ -60,6 +60,7 @@ dispatcher.add_handler(CommandHandler('help', help_handler))
 dispatcher.add_handler(CommandHandler('cookbook', cookbook_handler))
 dispatcher.add_handler(MessageHandler(Filters.reply, reply_handler))
 dispatcher.add_handler(MessageHandler(Filters.text, main_handler))
+dispatcher.add_handler(MessageHandler(Filters.all, lambda bot, update: print(update.message)))
 
 if environ.get('ENVIRONMENT', None) == 'HEROKU':
 	print("Starting Webhook")
